# 🚀 Progressive Implementation: Autonomous Features with Persistent Navigation & UI Fixes